+++ test/strategy/Setup.t.sol
@@ -13,7 +13,7 @@ import "../mocks/MockERC20.sol";
 contract TestBaseStrategy is Test {
     uint256 internal constant RISK_RATIO = 109544511;
 
-    uint256 internal constant WETH_ASSET_ID = 1;
+    uint64 internal constant WETH_ASSET_ID = 1;
 
     Controller internal controller;
     MockERC20 internal usdc;
